pyproject.toml
[tool.poetry]
name = "spec-assistant"
version = "0.1.0"
description = "RAG backend for technical spec PDF Q&A (coils/transformers)"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
python-multipart = "^0.0.9"
python-dotenv = "^1.0.1"
unstructured = {extras = ["all-docs"], version = "^0.17.2"}
pinecone-client = "^3.0.0"
langchain = "^0.1.17"
httpx = "^0.27.0"
pytest = "^8.2.0"
pdfminer-six = "^20250506"
pillow-heif = "^0.22.0"
openai = "^1.82.1"
docx2pdf = "^0.1.8"
pypandoc = "^1.15"
xlsx2html = "^0.6.2"
pdfkit = "^1.0.0"
pymupdf = "^1.26.0"
nltk = "^3.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
